Title: 코드트리 DB, Time: 2082ms, Memory: 3MB, Status: Time Limit Exceed - Codetree

+++ 250406/코드트리 DB/codetree-db.cpp
@@ -63,7 +63,7 @@ int main() {
         else if(cmd == "sum"){
             int k;
             cin >> k;
-            int output = 0;
+            long long output = 0;
             for(auto iter = DB_v.begin(); iter != DB_v.end(); iter++){
                 if(iter->first > k){
                     break;
